* FileList: touch only one file from a given list possible. The directory line does not contain a date up to now. * WikistyleTextToSimpleXml.java : new Now writes paragraphs in 2 or more  columns of a table   if "..............." is found on start of line. The end is "^^^^^^^^^^^^^^^^^^^^^^^^" .   Gardening of the algorithm. Line ending with \r\n or only \r are supported too, see {@link #endline(String)}

+++ org/vishia/util/FileSystem.java
@@ -559,13 +559,18 @@ public class FileSystem
    */
   public static boolean cleandir(File dir){
     boolean bOk = true;
-    if(!dir.isDirectory()){ dir = getDir(dir); }
-    File[] files = dir.listFiles();
-    for(File file: files){
-      if(file.isDirectory()){
-        bOk = bOk && rmdir(file);
-      } else {
-        bOk = bOk && file.delete();
+    if(!dir.exists()) {
+      bOk = dir.mkdirs();
+    }
+    if(bOk) {
+      if(!dir.isDirectory()){ dir = getDir(dir); }
+      File[] files = dir.listFiles();
+      for(File file: files){
+        if(file.isDirectory()){
+          bOk = bOk && rmdir(file);
+        } else {
+          bOk = bOk && file.delete();
+        }
       }
     }
     return bOk;
